backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/LLM
pydantic-ai==0.0.13
openai==1.3.0
anthropic==0.7.0

# PDF Processing
pymupdf==1.23.8
pytesseract==0.3.10
Pillow==10.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Authentication
fastapi-users[sqlalchemy]==12.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Azure SDK
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
testcontainers==3.7.1
playwright==1.40.0
